feat: Add function to add products to cart
Added a new function 'addCartClicked' to add selected products to the shopping cart. It fetches the product title, price, and image and adds it to the cart. Also, it updates the total price after adding the product. If the product is already in the cart, it triggers an alert message.